// src/endpoints/token/data/dtos/eos-dac.dtos.ts
/**
 * Represents the data structure of the eosDac
 * chain/get_info response
 * @type
 */
export type EosDacInfoDto = {
  server_version: string;
  chain_id: string;
  head_block_num: number;
  last_irreversible_block_num: number;
  last_irreversible_block_id: string;
  head_block_id: string;
  head_block_time: string;
  head_block_producer: string;
  virtual_block_cpu_limit: number;
  virtual_block_net_limit: number;
  block_cpu_limit: number;
  block_net_limit: number;
  server_version_string: string;
  fork_db_head_block_num: number;
  fork_db_head_block_id: string;
  server_full_version_string: string;
};

/**
 * Represents the data structure of the eosDac
 * chain/get_currency_stats response
 * @type
 */
export type CurrencyStatsDto = {
  TLM: {
    supply: string;
    max_supply: string;
    issuer: string;
  };
};